Convert MaxWasmSize to int

+++ x/wasm/types/validation.go
@@ -10,13 +10,13 @@ const (
 )
 
 // MaxWasmSize is the largest a compiled contract code can be when storing code on chain
-var MaxWasmSize uint64 = 800 * 1024
+var MaxWasmSize int = 800 * 1024
 
 func validateWasmCode(s []byte) error {
 	if len(s) == 0 {
 		return sdkerrors.Wrap(ErrEmpty, "is required")
 	}
-	if uint64(len(s)) > MaxWasmSize {
+	if len(s) > MaxWasmSize {
 		return sdkerrors.Wrapf(ErrLimit, "cannot be longer than %d bytes", MaxWasmSize)
 	}
 	return nil
